fix incorrect mod name

+++ output/font_alpha.rs
@@ -1,5 +1,5 @@
 
-pub const HEIGHT: usize = 22;
+pub const HEIGHT: u8 = 22;
 
 pub const SET: [&str; 127] = [
     "",
