+++ output/font_shimrod.rs
@@ -1,5 +1,5 @@
 
-pub const HEIGHT: usize = 6;
+pub const HEIGHT: u8 = 6;
 
 pub const SET: [&str; 127] = [
     "",
