+++ output/font_swan.rs
@@ -1,5 +1,5 @@
 
-pub const HEIGHT: usize = 9;
+pub const HEIGHT: u8 = 9;
 
 pub const SET: [&str; 127] = [
     "",
