+++ output/font_tsalagi.rs
@@ -1,5 +1,5 @@
 
-pub const HEIGHT: usize = 5;
+pub const HEIGHT: u8 = 5;
 
 pub const SET: [&str; 127] = [
     "",
