+++ src/code_template.rs
@@ -1,7 +1,7 @@
 pub const PAT: &str = "{$}";
 
 pub const HEIGHT_TEMP: &str = r#"
-pub const HEIGHT: usize = {$};
+pub const HEIGHT: u8 = {$};
 "#;
 
 pub const ARR_TEMP: &str = r#"
